Update Alt+Q accelerator

+++ ts/accelerators.ts
@@ -3,6 +3,7 @@ import { getAppById, spawnOverlay } from "$ts/apps";
 import { Process } from "$ts/process";
 import { GlobalDispatch } from "$ts/process/dispatch/global";
 import { focusedPid } from "$ts/stores/apps";
+import { sleep } from "$ts/util";
 import { AppKeyCombinations } from "$types/accelerator";
 import { ActionCenterOpened, StartMenuOpened } from "./stores";
 
@@ -24,16 +25,24 @@ export const ShellAccelerators: (process: Process) => AppKeyCombinations = (proc
     alt: true,
     key: "q",
     global: true,
-    action() {
+    async action() {
       const pid = focusedPid.get();
 
       if (!pid) return;
 
       const proc = process.handler.getProcess(pid)
 
+      if (!proc) return;
+
+      const parentPid = proc.parentPid
+
       if (proc.app && proc.app.metadata.core) return;
 
-      process.handler.kill(pid)
+      process.handler.kill(pid);
+
+      await sleep()
+
+      if (parentPid) focusedPid.set(parentPid)
     },
   },
   {
